+++ frontend/srcs/src/app/guards/login.guard.ts
@@ -1,8 +1,8 @@
 import { Injectable } from '@angular/core';
-import { Router } from '@angular/router';
 import { AuthService } from '../services/auth.service';
-import { SessionStorageConstants } from '../utils/session.storage';
+import { CookieConstants } from '../utils/cookie.constants';
 import { environment } from 'src/environments/environment';
+import { UsersCache } from '../cache';
 
 
 @Injectable({
@@ -10,14 +10,17 @@ import { environment } from 'src/environments/environment';
 })
 export class LoginGuard {
   constructor(
-    private readonly router: Router,
+    private readonly cachedUsers: UsersCache,
     private auth: AuthService
   ){}
   canActivate(): boolean {
-    const checkSession = this.auth.readFromCookie(SessionStorageConstants.USER_TOKEN)
+    const checkSession = this.auth.readFromCookie(CookieConstants.USER_TOKEN)
     if (checkSession.sub === ''){
       window.location.href = environment.loginUrl;
       return false;
+    }
+    if (this.cachedUsers.getMyUserId()!== checkSession.sub) {
+      this.cachedUsers.setMyUserId(checkSession.sub);
     }
     return true;
   }