+++ frontend/srcs/src/app/services/web-socket.service.ts
@@ -2,7 +2,7 @@
 import { Socket } from 'ngx-socket-io';
 import { environment } from 'src/environments/environment';
 import { CookieService } from 'ngx-cookie-service';
-import { SessionStorageConstants } from '../utils';
+import { CookieConstants } from '../utils';
 import { Injectable } from '@angular/core';
 
 @Injectable()
@@ -11,6 +11,6 @@ export class MySocket extends Socket {
         private cookieSevice: CookieService
     ) {
         super({ url: environment.apiUrl, options: {} });
-        this.ioSocket['auth'] = { token: this.cookieSevice.get(SessionStorageConstants.USER_TOKEN) };
+        this.ioSocket['auth'] = { token: this.cookieSevice.get(CookieConstants.USER_TOKEN) };
     }
 }