+++ frontend/srcs/src/app/utils/cookie.constants.ts
@@ -1,4 +1,4 @@
-export class SessionStorageConstants {
+export class CookieConstants {
   public static readonly USER_TOKEN = 'USER_TOKEN';
   // public static readonly MENUS = 'MENUS';
   // public static readonly USER_REFRESH_TOKEN = 'USER_REFRESH_TOKEN';