+++ frontend/srcs/src/app/utils/index.ts
@@ -1,3 +1,3 @@
 export * from './uris.contants';
-export * from './session.storage';
+export * from './cookie.constants';
 export * from './date.mutations'