+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 2023-10-31 - emadriga
+### Changed
+- session.storage.ts remane into `cookie.constants.ts`
+- cache currentClientUserId to avoid calcs
+
 ## 2023-10-30 - emadriga
 ### Added
 - Added front `PATCH` & `PUT` calls